changed how user roles are granted.
now the manager makes addresses users in a separate function, and the deposit functions are restricted to be callable by only USERs. this makes more sense than when, previously, we made the manager handle the depositing as well. This places control of depositing user funds in the manager's hands, which is undesirable. Now, once an address is granted USER access, they can call the deposit functions themselves.